requirements.txt
altgraph==0.15
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
html5lib==1.0.1
idna==2.7
macholib==1.9
pefile==2017.11.5
Pillow==5.2.0
PyInstaller==3.3.1
pymodm==0.4.0
pymongo==3.7.1
pypiwin32==223
pywin32==223
requests==2.19.1
six==1.11.0
urllib3==1.23
webencodings==0.5.1
XlsxWriter==1.0.5
